upping

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -5,17 +5,17 @@
 */
 int main(void)
 {
-	long long i, a = 1, b = 2, tmp;
+	long i, a = 1, b = 2, tmp;
 
-	printf("%lld, %lld, ", a, b);
+	printf("%ld, %ld, ", a, b);
 	for (i = 0; i < 48; i++)
 	{
 		tmp = a + b;
 		a = b;
 		b = tmp;
 		if (i == 47)
-			printf("%lld", b);
-		printf("%lld, ", b);
+			printf("%ld", b);
+		printf("%ld, ", b);
 	}
 	putchar('\n');
 	return (0);
